refactor(clang-tidy): Address critical bugprone and implicit conversion warnings
- Fixed bugprone-narrowing-conversions by adding explicit static_casts.
- Made all readability-implicit-bool-conversions explicit by comparing pointers and size_t results to nullptr and 0 respectively.

This commit addresses the highest-priority static analysis warnings to improve code safety and prevent potential bugs.

+++ VerticalSlice/src/AttackGem.cpp
@@ -4,9 +4,9 @@
 #include "Monster.h"
 
 float AttackGem::onMatch(Board& board, Player& player, Monster& monster) {
-    float baseDamage = 1.0f + (player.getVigor() / 10.0f); // Every 10 vigor adds 1 base damage
+    float baseDamage = 1.0f + (static_cast<float>(player.getVigor()) / 10.0f); // Every 10 vigor adds 1 base damage
     float multiplier = player.getStatModifier("skull_damage");
-    float damage = (baseDamage + getLevel() - 1) * multiplier;
+    float damage = (baseDamage + static_cast<float>(getLevel()) - 1) * multiplier;
     std::cout << "[ATTACK GEM] Matched Skull Gem (Level " << getLevel() << ") dealt " << damage << " damage." << std::endl;
     return damage; // Add level bonus
 }
